requirements.txt
azure-core==1.31.0
azure-messaging-webpubsubservice==1.2.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
